Separated API logic + Added PostFirewallPolicy
