+++ main.go
@@ -17,21 +17,30 @@ type FirewallRuleCache struct {
 	groupId string
 }
 
+type FirewallZonePolicyCache struct {
+	id        string
+	groupId   string
+	srcZoneId string
+	dstZoneId string
+}
+
 type ZoneCache struct {
 	id string
 }
 
 type unifiAddrList struct {
-	c                  unifi.Client
-	cacheIpv4          map[string]bool
-	cacheIpv6          map[string]bool
-	firewallGroupsIPv4 map[string]string
-	firewallGroupsIPv6 map[string]string
-	firewallRuleIPv4   map[string]FirewallRuleCache
-	firewallRuleIPv6   map[string]FirewallRuleCache
-	modified           bool
-	isZoneBased        bool
-	firewallZones      map[string]ZoneCache
+	c                     unifi.Client
+	cacheIpv4             map[string]bool
+	cacheIpv6             map[string]bool
+	firewallGroupsIPv4    map[string]string
+	firewallGroupsIPv6    map[string]string
+	firewallRuleIPv4      map[string]FirewallRuleCache
+	firewallRuleIPv6      map[string]FirewallRuleCache
+	firewallZonePoliyIPv4 map[string]FirewallZonePolicyCache
+	firewallZonePoliyIPv6 map[string]FirewallZonePolicyCache
+	modified              bool
+	isZoneBased           bool
+	firewallZones         map[string]ZoneCache
 }
 
 // This variable is set by the build process with ldflags
